Add unit tests and webcam demo for BlazeFace detection
- Introduced test data CSV for unit testing with various bounding box annotations.
- Implemented a test suite in `run_tests.py` to discover and run unit tests.
- Created unit tests for the collate function in `test_collate.py` to validate box padding and handling of zero boxes.
- Developed comprehensive tests in `test_pipeline.py` for dataset loading, normalization, resizing, anchor encoding, and decoding.
- Added a webcam demo script in `webcam_demo.py` for real-time BlazeFace detection with configurable parameters.

+++ loss_functions.py
@@ -36,7 +36,7 @@ class BlazeFaceDetectionLoss(nn.Module):
     
     def __init__(
         self,
-        hard_negative_ratio: int = 1,
+        hard_negative_ratio: float = 1.0,
         detection_weight: float = 150.0,
         classification_weight: float = 35.0,
         scale: int = 128,
@@ -315,7 +315,7 @@ def get_loss(**kwargs) -> nn.Module:
 
     Args:
         **kwargs: Arguments for BlazeFaceDetectionLoss
-            - hard_negative_ratio: Ratio of negatives to positives (default: 1)
+            - hard_negative_ratio: Ratio of negatives to positives (default: 1.0)
             - detection_weight: Weight for box regression (default: 150.0)
             - classification_weight: Weight for classification (default: 35.0)
             - scale: Image scale for decoding (default: 128)
